to-Do-List
Creating a To-Do-List using python programming language. simple To-Do-List contains addition , deletion of task that is to be done or been completed. the to-do-list is build by importing tkinter module in the python file. the task list is text that is created to store the task that is to be created for temporary use.